pyproject.toml
[project]
name = "prashna-ai"
version = "1.0.0"
description = "Prashna-AI: Adaptive AI-Powered Quiz & Question Generator"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.6.0",
    "supabase>=2.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "pdfplumber>=0.10.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "spacy>=3.7.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
ml = [
    "sentence-transformers>=2.2.0",
    "bertopic>=0.15.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
]

[tool.setuptools]
packages = ["backend", "backend.routes", "backend.services"]
